requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
filetype==1.2.0
frozenlist==1.6.0
google-generativeai==0.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Js2Py_3.13==0.74.1
lxml==5.4.0
m3u8==6.0.0
multidict==6.4.3
outcome==1.3.0.post0
propcache==0.3.1
psycopg==3.2.7
psycopg-binary==3.2.7
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyjsparser==2.7.1
pyotp==2.9.0
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
twikit==2.3.3
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websocket-client==1.8.0
webvtt-py==0.5.1
wsproto==1.2.0
yarl==1.20.0
